webhook/requirements.txt
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#    https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

flask==3.1.0
functions-framework==3.8.2
google-cloud-aiplatform==1.87.0
google-cloud-documentai==3.3.0
google-cloud-firestore==2.20.1
google-cloud-storage==2.19.0
google-genai==1.9.0
